added enzyme and puppeteer unit tests
