+++ client/src/components/container.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Image from './image.jsx';
 import Info from './info.jsx';
 
-var WhatIsThis = ({ restaurant, onLeftClick, onRightClick, index }) => {
+var Container = ({ restaurant, onLeftClick, onRightClick, index }) => {
   return (
     <div className="container">
       <div className="wrapper">
@@ -18,4 +18,4 @@ var WhatIsThis = ({ restaurant, onLeftClick, onRightClick, index }) => {
   );
 };
 
-export default WhatIsThis;
+export default Container;
