+++ package.json
@@ -31,6 +31,9 @@
   },
   "devDependencies": {
     "axios": "^0.18.0",
+    "babel-jest": "^23.6.0",
+    "enzyme": "^3.6.0",
+    "enzyme-adapter-react-16": "^1.5.0",
     "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
@@ -41,6 +44,8 @@
     "eslint-plugin-react": "^7.11.1",
     "jest": "^23.6.0",
     "prettier": "^1.14.2",
-    "puppeteer": "^1.8.0"
+    "puppeteer": "^1.8.0",
+    "regenerator-runtime": "^0.12.1",
+    "sinon": "^6.3.1"
   }
 }
